% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin.R
\name{addPackageToFunction}
\alias{addPackageToFunction}
\title{Add Package Name to Function Calls}
\usage{
addPackageToFunction(
  pkg,
  file = NULL,
  overwrite = FALSE,
  ignoreComments = TRUE,
  verbose = FALSE,
  functions = NULL
)
}
\arguments{
\item{pkg}{a package name}

\item{file}{a path to a script}

\item{overwrite}{a boolean, if TRUE the file will be saved and overwritten. If FALSE the file is returned.}

\item{ignoreComments}{a boolean, if TRUE lines starting with # are ignored}

\item{verbose}{a boolean}

\item{functions}{a vector with function names}
}
\value{

}
\description{
Add Package Name to Function Calls
}
