% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functionality.R
\name{checkFunctions}
\alias{checkFunctions}
\title{Extract relevant Function information}
\usage{
checkFunctions(
  script,
  functions,
  ignoreComments = TRUE,
  pkg = NULL,
  verbose = TRUE
)
}
\arguments{
\item{script}{a script to check}

\item{functions}{a vector with function names}

\item{ignoreComments}{a boolean, if TRUE lines starting with # are ignored}
}
\value{

}
\description{
Extract relevant Function information
}
