% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_functions.R
\name{find_functions}
\alias{find_functions}
\title{Find function names}
\usage{
find_functions(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a character vector
}
\description{
Find function names
}
