% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exported_functions.R
\name{get_exported_functions}
\alias{get_exported_functions}
\title{Get All Exported Functions From a Package}
\usage{
get_exported_functions(pkg)
}
\arguments{
\item{pkg}{a character string of a package name}
}
\value{
character vector of functions names
}
\description{
Get All Exported Functions From a Package
}
\examples{
get_exported_functions("base")
}
