% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parsed_data.R
\name{get_parsed_data}
\alias{get_parsed_data}
\title{Retrieve Parse Information from a string}
\usage{
get_parsed_data(file = "", text = NULL)
}
\arguments{
\item{x}{an R script as a character vector}
}
\value{
data.frame
}
\description{
Retrieve Parse Information from a string
}
