% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functionality.R
\name{getFunctions}
\alias{getFunctions}
\title{Get All Exported Functions From a Package}
\usage{
getFunctions(pkg)
}
\arguments{
\item{pkg}{`character(1)`
Name of package}
}
\value{
`character(1)`
}
\description{
Get All Exported Functions From a Package
}
\examples{
getFunctions("data.table")

}
