% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorder_registry.R
\name{origin_recorders}
\alias{origin_recorders}
\title{Names of the registered recorders}
\usage{
origin_recorders()
}
\value{
A character vector, in registration order.
}
\description{
Names of the registered recorders
}
