% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originize.R
\name{originize}
\alias{originize}
\title{Add Explicit Package Names to its Functions}
\usage{
originize(
  script,
  file,
  functions,
  pkgs = .packages(),
  overwrite = FALSE,
  ignoreComments = TRUE,
  verbose = FALSE,
  html = TRUE
)
}
\arguments{
\item{file}{a path to a script}

\item{pkgs}{a vector with package names}

\item{overwrite}{a boolean, if TRUE the file will be saved and overwritten. If FALSE the file is returned.}

\item{ignoreComments}{a boolean, if TRUE lines starting with # are ignored}

\item{verbose}{a boolean}
}
\value{

}
\description{
Add Explicit Package Names to its Functions
}
