% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physicell_lineage.R
\name{palincode_static_ids}
\alias{palincode_static_ids}
\title{Draw unique static nucleotide identifiers for integrations}
\usage{
palincode_static_ids(number, identifier_length)
}
\arguments{
\item{number}{Number of identifiers to draw; zero returns an empty vector.}

\item{identifier_length}{Length in bases of each identifier.}
}
\value{
A character vector of `number` unique A/C/G/T strings.
}
\description{
Resamples until no two identifiers collide, after checking that
`4^identifier_length` can accommodate `number` distinct strings.
}
