% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logging.R
\name{prep_verbose}
\alias{prep_verbose}
\title{Print Change Log and potential Errors in Console}
\usage{
prep_verbose(
  script_prior,
  lineMatches = rep(TRUE, length(script_after)),
  functions,
  functionsInScript,
  special_functions = NULL,
  special_matches = FALSE
)
}
\arguments{
\item{script_prior}{the script before any changes}

\item{lineMatches}{a boolean vector with the lines that contain changes}

\item{functions}{a vector with function names}

\item{functionsInScript}{a vector which functions were used}

\item{special_functions}{a vector with special functions such as \%like\%}

\item{special_matches}{a boolean vector that indicates which special
functions are used}

\item{script_after}{the script after the changes}
}
\value{

}
\description{
Print Change Log and potential Errors in Console
}
