% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_extract.R
\name{reg_extract}
\alias{reg_extract}
\title{Extract string by regular expression}
\usage{
reg_extract(x, pattern, invert = FALSE, ...)
}
\arguments{
\item{x}{a character vector}

\item{pattern}{a regular expression}

\item{invert}{logical: if TRUE, extract or replace the non-matched substrings.}

\item{...}{arguments passed to \link[base]{gregexpr}}
}
\value{
a character vector
}
\description{
Extract string by regular expression
}
