% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/un_list.R
\name{un_list}
\alias{un_list}
\title{Unlist a list into a vector with names equal to former list element name}
\usage{
un_list(l, nms = names(l))
}
\arguments{
\item{l}{a list to convert into a vector}

\item{nms}{character vector of same length as list, defaults to the
names of l}
}
\value{
named vector
}
\description{
The regular `base::unlist()` function creates a named vector in which each
element has a name corresponding to the list element it belonged to priorly
plus a number to yield unique names. This function basically excludes the
number part
}
\details{
the `base::unlist` function converts a list into a vector yet
   assigns unique names to each vector element, More precisely, it adds
    a number to the name of its list element. This function does not
    create unique names but assigns the bare name of the list element
    to all vector elements that stem from this list element
}
\examples{
l <- list(red = 1:3, blue = 1:2)
un_list(l)
# >  red  red  red blue blue
# >    1    2    3    1    2
}
