% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_functions.R
\name{get_local_functions}
\alias{get_local_functions}
\title{Find all user defined functions in the project}
\usage{
get_local_functions(path = ".")
}
\arguments{
\item{path}{Path in which all defined function names should be found and retrieved}
}
\value{
character vector of function names
}
\description{
Find all user defined functions in the project
}
\examples{
get_local_functions(path = ".")
get_local_functions(path = rprojroot::find_rstudio_root_file())
}
