% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkgs_from_description.R
\name{get_pkgs_from_description}
\alias{get_pkgs_from_description}
\title{Get Packages from the DESCRIPTION file}
\usage{
get_pkgs_from_description()
}
\value{
character vector of package names
}
\description{
Get Packages from the DESCRIPTION file
}
\examples{
\dontrun{
get_pkgs_from_description()
}
}
