\name{origin-package}
\alias{origin-package}
\docType{package}
\title{
Parent of origin testing
}
\description{
Parent of origin testing using the method proposed by Weinberg, with extension to accomodate more than one affeected offspring in a family.
}
\details{
\tabular{ll}{
Package: \tab origin\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-12-09\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}

This is a small package, with one main function, \code{origin},
and a <code>print</code> method defined.  See \code{help(origin)}
for details.
}
\author{
Chris Wallace <chris.wallace@cimr.cam.ac.uk>
}
\examples{
## origin(genotype,data=mydataframe)
}
