% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin.R
\docType{package}
\name{origin-package}
\alias{origin}
\alias{origin-package}
\title{origin: Explicitly Qualifying Namespaces by Automatically Adding 'pkg::' to Functions}
\description{
To learn more about origin, start with the vignette:
`browseVignettes(package = "origin")`
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mnist91/origin}
  \item Report bugs at \url{https://github.com/mnist91/origin/issues}
}

}
\keyword{internal}
