% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originize_file.R
\name{originize_file}
\alias{originize_file}
\title{Add Explicit Package Names to its Functions}
\usage{
originize_file(
  file,
  pkgs = .packages(),
  overwrite = FALSE,
  ignoreComments = TRUE,
  excludeBasePackages = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{file}{a path to a script}

\item{pkgs}{a vector with package names}

\item{overwrite}{a boolean, if TRUE the file will be saved and overwritten. If FALSE the file is returned.}

\item{ignoreComments}{a boolean, if TRUE lines starting with # are ignored}

\item{excludeBasePackages}{a boolean, if TRUE base R functions are excluded}

\item{verbose}{a boolean}
}
\value{

}
\description{
Add Explicit Package Names to its Functions
}
