% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originize_selection.R
\name{originize_selection}
\alias{originize_selection}
\title{Add Explicit Package Names to its Functions}
\usage{
originize_selection(
  context = rstudioapi::getSourceEditorContext(),
  pkgs = .packages(),
  overwrite = TRUE,
  ask_before_applying_changes = TRUE,
  ignoreComments = TRUE,
  check_conflicts = TRUE,
  check_base_conflicts = TRUE,
  add_base_packages = FALSE,
  excluded_functions = list(),
  verbose = TRUE,
  html = TRUE
)
}
\arguments{
\item{pkgs}{a vector with package names}

\item{overwrite}{a boolean, if TRUE the file will be saved and overwritten. If FALSE the file is returned.}

\item{ignoreComments}{a boolean, if TRUE lines starting with # are ignored}

\item{verbose}{a boolean}

\item{file}{a path to a script}

\item{excludeBasePackages}{a boolean, if TRUE base R functions are excluded}
}
\value{

}
\description{
Add Explicit Package Names to its Functions
}
