% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originize_selection.R
\name{originize_selection}
\alias{originize_selection}
\title{Add Explicit Package Names to its Functions}
\usage{
originize_selection(
  context = rstudioapi::getSourceEditorContext(),
  pkgs = getOption("origin.pkgs"),
  overwrite = TRUE,
  ask_before_applying_changes = getOption("origin.ask_before_applying_changes"),
  ignore_comments = TRUE,
  check_conflicts = TRUE,
  check_base_conflicts = TRUE,
  add_base_packages = FALSE,
  excluded_functions = list(),
  verbose = TRUE,
  use_markers = getOption("origin.use_markers_for_logging")
)
}
\arguments{
\item{pkgs}{a vector with package names}

\item{overwrite}{a boolean, if TRUE the file will be saved and overwritten.
If FALSE the file is returned.}

\item{ignore_comments}{a boolean, if TRUE lines starting with # are ignored}

\item{verbose}{a boolean}

\item{file}{a path to a script}

\item{excludeBasePackages}{a boolean, if TRUE base R functions are excluded}
}
\value{

}
\description{
Add Explicit Package Names to its Functions
}
