% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pkg_usage.R
\name{print.pkg_usage}
\alias{print.pkg_usage}
\title{Print the summary of check_pkg_usage}
\usage{
\method{print}{pkg_usage}(x, ...)
}
\arguments{
\item{x}{a pkg_usage_object}

\item{...}{passed to other methods}
}
\value{
x invisibly
}
\description{
Print the summary of check_pkg_usage
}
\examples{
\dontrun{
result <- check_pkg_usage()
print(result)
}
}
