% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functionality.R
\name{un_list}
\alias{un_list}
\title{Unlist a list into a vector with names equal to fromer list element name}
\usage{
un_list(l, nms = names(l))
}
\arguments{
\item{l}{a list to convert into a vector}

\item{nms}{character vector of same length as list, defaults to the 
names of l}
}
\value{
named vector
}
\description{
Unlist a list into a vector with names equal to fromer list element name
}
\details{
the `base::unlist` function converts a list into a vector yet
   assigns unique names to each vector element, More precisely, it adds
    a number to the name of its list element. This function does not
    create unique names but assigns the bare name of the list element
    to all vector elements that stem from this list element
}
\examples{
l <- list(rot = 1:3, blau = 1:2)
un_list(l)
# >  rot  rot  rot blau blau
# >    1    2    3    1    2
}
